refactor: move instructor profile update function fron instructorsController.ts to adminsController.ts
