+++ backend/src/routes/adminsRouter.ts
@@ -5,6 +5,7 @@ import {
   registerPlanController,
   registerEventController,
   updateAdminProfileController,
+  updateInstructorProfileController,
   updateEventProfileController,
   deleteAdminController,
   deleteEventController,
@@ -45,12 +46,23 @@ adminsRouter.post(
   upload.single("icon"),
   registerInstructorController,
 );
+adminsRouter.patch(
+  "/instructor-list/update/:id",
+  verifyAuthentication,
+  upload.none(),
+  updateInstructorProfileController,
+);
+adminsRouter.patch(
+  "/instructor-list/update/:id/withIcon",
+  verifyAuthentication,
+  upload.single("icon"),
+  updateInstructorProfileController,
+);
 adminsRouter.post(
   "/plan-list/register",
   verifyAuthentication,
   registerPlanController,
 );
-
 adminsRouter.post(
   "/event-list/register",
   verifyAuthentication,
