+++ backend/src/routes/instructorsRouter.ts
@@ -10,7 +10,6 @@ import {
   getAllInstructorsController,
   getRecurringAvailabilityById,
   getInstructorProfileController,
-  updateInstructorProfile,
   getCalendarClassesController,
   getInstructorProfilesController,
   getSameDateClassesController,
@@ -37,7 +36,6 @@ import {
   addInstructorAbsenceController,
   removeInstructorAbsenceController,
 } from "../../src/controllers/instructorAbsenceController";
-import upload from "../middlewares/upload.middleware";
 
 export const instructorsRouter = express.Router();
 
@@ -50,18 +48,6 @@ instructorsRouter.get("/profiles", getInstructorProfilesController);
 instructorsRouter.get("/:id", getInstructor);
 instructorsRouter.get("/:id/profile", parseId, (req, res) =>
   getInstructorProfileController(req as RequestWithId, res),
-);
-instructorsRouter.patch(
-  "/:id",
-  verifyAuthentication,
-  upload.none(),
-  updateInstructorProfile,
-);
-instructorsRouter.patch(
-  "/:id/withIcon",
-  verifyAuthentication,
-  upload.single("icon"),
-  updateInstructorProfile,
 );
 instructorsRouter.get("/:id/recurringAvailability", parseId, (req, res) =>
   RecurringAvailability.get(req as RequestWithId, res),
