+++ frontend/src/app/helper/data/data.ts
@@ -1,5 +1,6 @@
 export const holidayEventId = 2;
 export const defaultEventIds = [1, 2, 3, 4];
+export const defaultUserImageUrl = "/images/default-user-icon.jpg";
 
 export const prefectures = [
   "海外 / Overseas",
